API

Public

Private

Types

Functions

GraphCombinations — Module

GraphCombinations.jl

GraphCombinations.jl is a package for the generation of graphs for a given set of different vertices. Given a set of valent vertices (vertices of degree k), the package generates all possible graphs that can be constructed with these vertices.

The implementation is rather naive and I am sure it can be significiantly improved. Any contributions or suggestions are welcome :)

This package is heavily inspired by this StackExachange post by AccidentalFourierTransform. A mathematica notebook using his code can be found in the examples folder.

Public

GraphCombinations.allgraphs — Function

allgraphs(
    n::Vector{Int64};
    connected
) -> Vector{Tuple{Vector{Pair{Int64, Int64}}, Float64}}

Generates all unique topologies for a given vertex specification. It takes in a vector where n[k] is the number of vertices of degree k. By default, it generates connected graphs, unconnected can be computed by setting the keyword argument connected=false.

Returns a Vector of Tuples (Vector{Edge}, S) with Edge a Tuple{Int64,Int64} representing an edge in the graph and S::Float64 the corresponding symmetry factor.

Example

For input n = [2, 0, 0, 2] (2 vertices of degree 1, 2 vertex of degree 4):

julia> using GraphCombinations

julia> allgraphs([2, 0, 0, 2])
3-element Vector{Tuple{Vector{Pair{Int64, Int64}}, Float64}}:
 ([1 => 3, 2 => 3, 3 => 4, 3 => 4, 4 => 4], 4.0)
 ([1 => 3, 2 => 4, 3 => 3, 3 => 4, 4 => 4], 4.0)
 ([1 => 3, 2 => 4, 3 => 4, 3 => 4, 3 => 4], 6.0)

GraphCombinations.build_graph — Function

build_graph(
    graph_rep::Vector{Pair{Int64, Int64}},
    num_vertices::Int64
) -> GraphCombinations.MultigraphWrap{Int64}

Builds a MultigraphWrap from an edges list.

GraphCombinations.combinatoric_factor — Function

combinatoric_factor(n) -> Any

Computes the combinatorial factor for graph symmetry calculations.

This function calculates the product of two factors:

The vertex permutation factor: number of ways to permute identical internal vertices

The edge endpoint permutation factor: number of ways to permute edge endpoints attached to vertices of the same degree

This factor is used in the calculation of symmetry factors for Feynman diagrams.

Parameters

n: Vector where n[k] is the number of vertices with degree k

Returns

A floating point number representing the total combinatorial factor

GraphCombinations.total_degree — Function

total_degree(n::Vector{Int64}) -> Int64

Calculates the total degree of a graph topology represented by a vector of the degrees of vertices.

GraphCombinations.canonical_form — Function

canonical_form(graph::GraphRep, internal_indices::UnitRange{Int})::GraphRep

Finds the canonical representation of a graph under permutations of internal vertices. The canonical form is the lexicographically smallest graph representation achievable through permutation of internal_indices.

Private

Types

GraphCombinations.MultigraphWrap — Type

MultigraphWrap{T} <: AbstractGraph{T}

A wrapper for Multigraph that implements the AbstractGraph interface from Graphs.jl.

This wrapper is necessary because GraphMakie.jl only supports the Graphs.jl interface, but not Multigraphs.jl directly. See the discussion at https://github.com/MakieOrg/GraphMakie.jl/issues/52 for more details.

The wrapper provides all necessary methods to make a Multigraph compatible with functions expecting an AbstractGraph from Graphs.jl, especially for visualization purposes with GraphMakie.jl.

Functions

GraphCombinations.corr — Function

corr(points::Vector{Int})

Computes all possible pairings (Wick contractions) of the given points. Points are represented by integers. The function returns a list of "terms", where each term is a list of Edge (Pair{Int, Int}) objects.

Edges a => b always have a < b.

Examples

julia> import GraphCombinations as GC

julia> GC.corr([1,2,3,4])
3-element Vector{Vector{Pair{Int64, Int64}}}:
 [1 => 2, 3 => 4]
 [1 => 3, 2 => 4]
 [1 => 4, 2 => 3]

GraphCombinations.sort_graph_edges — Function

sort_graph_edges(graph::GraphRep)::GraphRep

Sorts edges within a graph representation canonically. Ensures a < b in each a => b and then sorts the vector of edges.

GraphCombinations.create_points — Function

create_points(n::Vector{Int64}) -> Vector{Int64}

Converts a vertex degree specification n to a flat array of vertex indices.

For a vector n where n[k] specifies the number of vertices with degree k, this function creates a vector where each vertex index appears a number of times equal to its degree. This representation is used for generating all possible pairings in the correlation function calculation.

Example

For input n = [2, 1] (2 vertices of degree 1, 1 vertex of degree 2):

julia> import GraphCombinations as GC

julia> GC.create_points([2, 1])
4-element Vector{Int64}:
 1
 2
 3
 3